Move `AddToManager` to dedicated `add.go` file
